# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "# Notice: The URL in `Example Request` is currently wrong. The URL\
    \ to use is: `https://api.sandbox.mobilepay.dk/pos/api/v10/...`\n## PoS  V10 API\n\
    \nThe API is split into 5 categories: \n* **Brands**: Look up brands or stores\
    \ based on brand \n* **PointOfSales**: Management of Point of Sales including\
    \ create, delete and querying. Note that a Point of Sale is not editable. \n*\
    \ **Refunds**: Everything needed in order to do a refund (create, cancel and look\
    \ up) \n* **Stores**: Look up store information \n* **Payments**: Everything needed\
    \ in order to initiate and complete a payment. \n\n#### Payment flows\nPayments\
    \ can be started in two ways: as an initiated payment that is immediatedly ready\
    \ for user approval or as a prepared payment that is not yet ready for user approval\
    \ and must be explicitly marked as ready. Once a payment is ready for user approval,\
    \ it is issued to the user. Once the user accepts the payment request, and the\
    \ payment amount has been reserved, the payment must be captured to complete the\
    \ payment.\n\nThe API thus supports the following two payment flows:\n* Initiate\
    \ payment (POST /payments) --> poll until reserved (GET /payments/{paymentId})\
    \ --> capture (POST /payments/{paymentId}/capture) \n* Prepare payment (POST /payments/prepare)\
    \ --> mark as ready (POST /payments/{paymentId}/ready) --> poll until reserved\
...
basePath: "/pos"
schemes:
...
consumes:
...
produces:
...
security: []
paths:
...
  /api/v10/pointofsales/{posId}:
    get:
      tags:
      - "PointOfSales"
      summary: "Lookup a Point of Sale"
      operationId: "GetPos"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "posId"
        in: "path"
        description: "PosId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "The Point of Sale"
          examples:
            application/json:
              posId: "bc2a3687-539a-44d2-9e92-0d746dc90089"
              merchantPosId: "My-Pos-1"
              storeId: "939f2b0d-152e-4000-8e04-2e30aa9faf7f"
              name: "Register 1"
              beaconId: "123456789123456"
              callback:
                destination: "https://example.com"
                type: "Url"
              supportedBeaconTypes:
              - "QR"
              - "BluetoothMP2"
          schema:
            $ref: "#/definitions/PosResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Point of Sale not found"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "PointOfSales"
    delete:
      tags:
      - "PointOfSales"
      summary: "Delete a Point of Sale"
      operationId: "DeletePos"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "posId"
        in: "path"
        description: "PosId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "Point of Sale deleted"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Point of Sale not found"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "PointOfSales"
  /api/v10/pointofsales:
    get:
      tags:
      - "PointOfSales"
      summary: "Lookup Point of Sales for a given filter"
      operationId: "GetPosWithFilter"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "storeId"
        in: "query"
        description: "Restricts Point of Sales returned to those that belong to this\
          \ Store"
        required: false
        type: "string"
        format: "uuid"
      - name: "beaconId"
        in: "query"
        description: "Restricts Point of Sales returned to the one with this beaconId"
        required: false
        type: "string"
      - name: "merchantPosId"
        in: "query"
        description: "Restricts Point of Sales returned to the one with this merchant\
          \ PoS identifer"
        required: false
        type: "string"
      - name: "active"
        in: "query"
        description: "Restricts Point of Sales returned to those that are active (i.e.,\
          \ not deleted)"
        required: false
        type: "boolean"
        default: true
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "A list of Point of Sales"
          examples:
            application/json:
              posIds:
              - "c0000a0f-68b8-4759-847b-08d5284c344c"
          schema:
            $ref: "#/definitions/PosIdsResponse"
        "204":
          description: "No Point of Sales found that match the filter"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "409":
          description: "Conflict"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "PointOfSales"
    post:
      tags:
      - "PointOfSales"
      summary: "Create a Point of Sale"
      operationId: "CreatePos"
      consumes:
      - "application/json-patch+json"
      - "application/json"
      - "text/json"
      - "application/*+json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "request"
        description: ""
        required: true
        schema:
          $ref: "#/definitions/CreatePosRequest"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      - name: "X-MobilePay-Idempotency-Key"
        in: "header"
        description: "Used to ensure retried calls are handled correctly"
        required: true
        type: "string"
      responses:
        "200":
          description: "Point of Sale created"
          examples:
            application/json:
              posId: "c0000a0f-68b8-4759-847b-08d5284c344c"
          schema:
            $ref: "#/definitions/CreatePosResponse"
        "400":
          description: "Bad Request. See error response for details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Not Found"
        "409":
          description: "Point of Sale could not be created. See error response for\
            \ details"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "PointOfSales"
  /api/v10/pointofsales/{posId}/checkin:
    get:
      tags:
      - "PointOfSales"
      summary: "Lookup the current MobilePay User information checked-in at a Point\
        \ of Sale"
      operationId: "GetCheckIn"
      consumes: []
      produces:
      - "application/json"
      parameters:
      - name: "posId"
        in: "path"
        description: "PosId"
        required: true
        type: "string"
        format: "uuid"
      - name: "Authorization"
        in: "header"
        description: "Integrator's Bearer Token"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-Id"
        in: "header"
        description: "Integrator's MobilePay Client Id (Must be a valid GUID)"
        required: true
        type: "string"
        format: "uuid"
      - name: "X-MobilePay-Client-System-Name"
        in: "header"
        description: "Integrator's Certified System Name"
        required: true
        type: "string"
      - name: "X-MobilePay-Client-System-Version"
        in: "header"
        description: "Integrator's Certified System Version"
        required: true
        type: "string"
      responses:
        "200":
          description: "MobilePay User info"
          examples:
            application/json:
              isUserCheckedIn: true
              loyaltyIds:
              - "123456"
              pollDelayInMs: 100
          schema:
            $ref: "#/definitions/GetCheckInResponse"
        "401":
          description: "Unauthorized"
        "403":
          description: "Forbidden"
        "404":
          description: "Point of Sale not found"
        "500":
          description: "Internal server error"
          examples:
            application/json:
              code: "1234"
              message: "Error message explaining the error that happened while processing\
                \ the request"
              correlationId: "ee216854-78ca-43c7-99f8-7ce6ee451019"
          schema:
            $ref: "#/definitions/ErrorResponse"
      x-swagger-router-controller: "PointOfSales"
...
definitions:
...
  ErrorResponse:
    type: "object"
    properties:
      code:
        type: "string"
        readOnly: true
      message:
        type: "string"
        readOnly: true
      correlationId:
        type: "string"
        readOnly: true
...
  PosResponse:
    type: "object"
    required:
    - "merchantPosId"
    - "name"
    - "posId"
    - "storeId"
    properties:
      posId:
        type: "string"
        format: "uuid"
      merchantPosId:
        type: "string"
        description: "Merchant defined Id for the given Point Of Sale"
      storeId:
        type: "string"
        format: "uuid"
      name:
        type: "string"
      beaconId:
        type: "string"
        description: "Former PosUnit"
      callback:
        $ref: "#/definitions/Callback"
      supportedBeaconTypes:
        type: "array"
        uniqueItems: true
        items:
          type: "string"
          enum:
          - "QR"
          - "NFC"
          - "BluetoothOther"
          - "BluetoothMP1"
          - "BluetoothMP2"
          - "BluetoothMP3"
          - "BluetoothMP4"
  Callback:
    type: "object"
    properties:
      destination:
        type: "string"
      type:
        type: "string"
        enum:
        - "Url"
        - "Alias"
  CreatePosRequest:
    type: "object"
    required:
    - "merchantPosId"
    - "name"
    - "storeId"
    - "supportedBeaconTypes"
    properties:
      merchantPosId:
        type: "string"
        description: "Merchant defined Id for the given Point Of Sale"
      storeId:
        type: "string"
        format: "uuid"
      name:
        type: "string"
      beaconId:
        type: "string"
        description: "Former PosUnitId"
      callback:
        $ref: "#/definitions/Callback"
      supportedBeaconTypes:
        type: "array"
        description: "List all the beacon type the Point of Sale supports for checkin:\
          \ QR, Bluetooth"
        uniqueItems: true
        items:
          type: "string"
          enum:
          - "QR"
          - "NFC"
          - "BluetoothOther"
          - "BluetoothMP1"
          - "BluetoothMP2"
          - "BluetoothMP3"
          - "BluetoothMP4"
    example:
      merchantPosId: "My-Pos-1"
      storeId: "939f2b0d-152e-4000-8e04-2e30aa9faf7f"
      name: "Register 1"
      beaconId: "123456789123456"
      callback:
        destination: "https://example.com"
        type: "Url"
      supportedBeaconTypes:
      - "QR"
      - "BluetoothMP2"
  CreatePosResponse:
    type: "object"
    required:
    - "posId"
    properties:
      posId:
        type: "string"
        format: "uuid"
  PosIdsResponse:
    type: "object"
    required:
    - "posIds"
    properties:
      posIds:
        type: "array"
        items:
          type: "string"
          format: "uuid"
  GetCheckInResponse:
    type: "object"
    required:
    - "isUserCheckedIn"
    - "loyaltyIds"
    - "pollDelayInMs"
    properties:
      isUserCheckedIn:
        type: "boolean"
      loyaltyIds:
        type: "array"
        items:
          type: "string"
      pollDelayInMs:
        type: "integer"
        format: "int32"
...